Dev (#38)
* updated clang-format

* hopefully a fix for macintosh

* hopefully a fix for macintosh

* hopefully a fix for macintosh

* updated changelog & now fixed macintosh

* changed naming

* updated docs & coverage reporting

* Docs updated (#32)

* updated docs & coverage reporting

* fixed typo in coverage yml file

* fixed workflow action trigger

* fixed another typo in coverage yml file

* fixed another issue with coverage.yml

* added security workflow

+++ tools/build_documentation.ps1
@@ -1,8 +1,10 @@
-
 $currentLocation = (Get-Location | Out-String)
 
-Get-ChildItem -Path "./Dokky" -Recurse | Remove-Item -Force -Recurse -Confirm:$false
-Remove-Item "./Dokky" -Recurse -Force -Confirm:$false
+if((Test-Path "./Dokky")){
+	Write-Host "Removing old files"
+	Get-ChildItem -Path "./Dokky" -Recurse | Remove-Item -Force -Recurse -Confirm:$false
+	Remove-Item "./Dokky" -Recurse -Force -Confirm:$false
+}
 
 git clone https://github.com/james-d12/Dokky.git
 Copy-Item ./docgen.json -Destination "./Dokky"
